refactor: totalPrice 字段 double/Double → BigDecimal
- 改造 5 层 6 文件: Order/DeletedOrder/OrderDTO/OrderItem(chat)/OrderServiceImpl/OrderConverter
- 适配 6 个测试文件的类型变更 (类型转换 + 序列化格式)
- 补全非 API 测试: 新建 OrderConverterTest(18) + OrderModelTest(16)
- 5 个测试类新增边界用例 +29 (OrderServiceImplTest +12 / OrderEventConsumerTest +8
  / FileFallbackDaemonTest +5 / RedisOrderIdGeneratorTest +4 含 OID-01 记录)
- 201/201 全部测试通过 (order-service 179 + chat-service 22)

Refs: docs/superpowers/specs/2026-06-02-totalPrice-BigDecimal-design.md
Plan: docs/superpowers/plans/2026-06-02-totalPrice-BigDecimal-refactor.md

+++ AI-Shopping-backend_Eureka/chat-service/src/main/java/com/gzasc/aishopping/chat/dto/OrderItem.java
@@ -1,10 +1,12 @@
 package com.gzasc.aishopping.chat.dto;
 
+import java.math.BigDecimal;
+
 public record OrderItem(
     String orderId,
     String productId,
     Integer quantity,
-    Double totalPrice,
+    BigDecimal totalPrice,
     String orderStatus,
     String orderDate,
     String contactName,
